+++ src/hooks/init/chatHooks.js
@@ -5,7 +5,25 @@ export function chatHooks() {
     chatImg.click((event) => {
       event.preventDefault();
       event.stopPropagation();
-      new ImagePopout($(chatImg).attr("src"), { shareable: true }).render(true);
+      if (foundry.applications?.apps?.ImagePopout) {
+        new foundry.applications.apps.ImagePopout({
+          src: $(chatImg).attr("src"),
+          showTitle: false,
+        }).render(true);
+      } else {
+        new ImagePopout($(chatImg).attr("src"), { shareable: true }).render(true);
+      }
+    });
+
+    const settingsButton = html.find("button.ddb-importer-chat-settings");
+
+    settingsButton.click((event) => {
+      event.preventDefault();
+      const menu = game.settings.menus.get(event.currentTarget.dataset.key);
+      const app = new menu.type({
+        sheetTab: "core",
+      });
+      return app.render(true);
     });
   });
 }